requirements.txt
async-timeout==4.0.2
attrs==22.2.0
cffi==1.15.1
click==8.1.6
coverage==7.2.5
cryptography==41.0.2
dnspython==2.2.1
exceptiongroup==1.1.0
iniconfig==2.0.0
jsonschema==4.17.3
neo4j==5.10.0
networkx==3.1
numpy==1.24.1
packaging==23.0
pandas==2.0.3
pika==1.3.2
pluggy==1.0.0
pycparser==2.21
pymongo==4.3.3
pyrsistent==0.19.3
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.6.0
rq==1.15.1
ruff==0.0.265
six==1.16.0
tc-messageBroker==1.4.0
tomli==2.0.1
types-pyOpenSSL==23.2.0.2
tzdata==2023.3
